+++ arrow/src/array/builder/boolean_buffer_builder.rs
@@ -1,8 +1,25 @@
-use crate::util::bit_util;
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
 
 use crate::buffer::{Buffer, MutableBuffer};
 
-use crate::array::builder::Range;
+use super::Range;
+
+use crate::util::bit_util;
 
 #[derive(Debug)]
 pub struct BooleanBufferBuilder {
@@ -154,12 +171,6 @@ impl From<BooleanBufferBuilder> for Buffer {
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    // use crate::array::array::Array;
-    // use crate::array::builder::ArrayBuilder;
-    use crate::array::Int32BufferBuilder;
-    use crate::array::Int8Builder;
-    use crate::array::UInt8BufferBuilder;
 
     #[test]
     fn test_boolean_buffer_builder_write_bytes() {
