+++ arrow/src/array/builder/boolean_builder.rs
@@ -1,16 +1,31 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
 use std::any::Any;
 use std::sync::Arc;
 
-use crate::datatypes::*;
-
+use crate::array::ArrayBuilder;
+use crate::array::ArrayData;
+use crate::array::ArrayRef;
+use crate::array::BooleanArray;
+use crate::datatypes::DataType;
 use crate::error::{ArrowError, Result};
 
-pub use crate::array::ArrayBuilder;
-pub use crate::array::ArrayData;
-pub use crate::array::ArrayRef;
-pub use crate::array::BooleanArray;
-
-pub use crate::array::builder::boolean_buffer_builder::BooleanBufferBuilder;
+use super::BooleanBufferBuilder;
 
 ///  Array builder for fixed-width primitive types
 ///
@@ -159,11 +174,6 @@ impl ArrayBuilder for BooleanBuilder {
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    use crate::array::Array;
-    use crate::bitmap::Bitmap;
-    use crate::buffer::Buffer;
-    use crate::error::Result;
 
     #[test]
     fn test_boolean_array_builder_append_slice() {
