+++ arrow/src/array/builder/buffer_builder.rs
@@ -1,9 +1,26 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
 use std::mem;
 
-use crate::array::builder::PhantomData;
+use crate::buffer::{Buffer, MutableBuffer};
 use crate::datatypes::ArrowNativeType;
 
-use crate::buffer::{Buffer, MutableBuffer};
+use super::PhantomData;
 
 ///  Converts a `MutableBuffer` to a `BufferBuilder<T>`.
 ///
@@ -274,8 +291,6 @@ impl<T: ArrowNativeType> BufferBuilder<T> {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-
     use crate::array::array::Array;
     use crate::array::builder::ArrayBuilder;
     use crate::array::Int32BufferBuilder;
