+++ arrow/src/array/builder/fixed_size_list_builder.rs
@@ -1,18 +1,33 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
 use std::any::Any;
 use std::sync::Arc;
 
-use crate::datatypes::*;
+use crate::array::ArrayData;
+use crate::array::ArrayRef;
+use crate::array::FixedSizeListArray;
+use crate::array::Int32BufferBuilder;
+use crate::datatypes::DataType;
+use crate::datatypes::Field;
 use crate::error::Result;
 
-pub use crate::array::builder::BooleanBufferBuilder;
-pub use crate::array::builder::BufferBuilder;
-pub use crate::array::ArrayBuilder;
-pub use crate::array::ArrayData;
-pub use crate::array::ArrayRef;
-pub use crate::array::FixedSizeListArray;
-pub use crate::array::OffsetSizeTrait;
-
-pub use crate::array::Int32BufferBuilder;
+use super::ArrayBuilder;
+use super::BooleanBufferBuilder;
 
 ///  Array builder for `ListArray`
 #[derive(Debug)]
@@ -147,12 +162,8 @@ mod tests {
     use super::*;
 
     use crate::array::Array;
-    use crate::bitmap::Bitmap;
-    use crate::buffer::Buffer;
-    use crate::error::Result;
-
-    use crate::array::builder::FixedSizeBinaryArray;
-    use crate::array::builder::FixedSizeBinaryBuilder;
+    use crate::array::FixedSizeBinaryArray;
+    use crate::array::FixedSizeBinaryBuilder;
     use crate::array::Int32Array;
     use crate::array::Int32Builder;
