+++ arrow/src/array/builder/generic_list_builder.rs
@@ -1,17 +1,32 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
 use std::any::Any;
 use std::sync::Arc;
 
-use crate::datatypes::*;
+use crate::array::ArrayData;
+use crate::array::ArrayRef;
+use crate::array::GenericListArray;
+use crate::array::OffsetSizeTrait;
+use crate::datatypes::DataType;
+use crate::datatypes::Field;
 use crate::error::Result;
 
-pub use crate::array::builder::BooleanBufferBuilder;
-pub use crate::array::builder::BufferBuilder;
-pub use crate::array::ArrayBuilder;
-pub use crate::array::ArrayData;
-pub use crate::array::ArrayRef;
-pub use crate::array::GenericListArray;
-pub use crate::array::OffsetSizeTrait;
-pub use crate::array::UInt8Builder;
+use super::{ArrayBuilder, BooleanBufferBuilder, BufferBuilder};
 
 ///  Array builder for `ListArray`
 #[derive(Debug)]
@@ -144,14 +159,12 @@ mod tests {
     use super::*;
 
     use crate::array::Array;
-    use crate::array::{Int32Array, Int32Builder};
-    use crate::bitmap::Bitmap;
+    use crate::array::Int32Array;
+    use crate::array::Int32Builder;
     use crate::buffer::Buffer;
-    use crate::error::Result;
 
-    pub use crate::array::builder::{
-        BinaryBuilder, GenericStringBuilder, LargeBinaryBuilder, LargeListBuilder,
-        LargeStringBuilder, ListBuilder, StringBuilder,
+    use crate::array::builder::{
+        BinaryBuilder, LargeBinaryBuilder, LargeListBuilder, ListBuilder, StringBuilder,
     };
 
     #[test]
