+++ arrow/src/array/builder/map_builder.rs
@@ -1,9 +1,33 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
 use std::any::Any;
 use std::sync::Arc;
 
-use crate::array::*;
-use crate::datatypes::*;
+use crate::array::array::Array;
+use crate::array::ArrayData;
+use crate::array::ArrayRef;
+use crate::array::MapArray;
+use crate::array::StructArray;
+use crate::datatypes::DataType;
+use crate::datatypes::Field;
 use crate::error::{ArrowError, Result};
+
+use super::{ArrayBuilder, BooleanBufferBuilder, BufferBuilder};
 
 #[derive(Debug)]
 pub struct MapBuilder<K: ArrayBuilder, V: ArrayBuilder> {
@@ -168,10 +192,10 @@ impl<K: ArrayBuilder, V: ArrayBuilder> ArrayBuilder for MapBuilder<K, V> {
 mod tests {
     use super::*;
 
-    use crate::array::Array;
+    use crate::array::builder::StringBuilder;
+    use crate::array::Int32Builder;
     use crate::bitmap::Bitmap;
     use crate::buffer::Buffer;
-    use crate::error::Result;
 
     // TODO: add a test that finishes building, after designing a spec-compliant
     // way of inserting values to the map.
