+++ arrow/src/array/builder/primitive_dictionary_builder.rs
@@ -1,10 +1,33 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
 use std::any::Any;
 use std::collections::HashMap;
 use std::sync::Arc;
 
-use crate::array::*;
-use crate::datatypes::*;
+use crate::array::ArrayRef;
+use crate::array::ArrowPrimitiveType;
+use crate::array::DictionaryArray;
+use crate::datatypes::ArrowNativeType;
+use crate::datatypes::ToByteSlice;
 use crate::error::{ArrowError, Result};
+
+use super::ArrayBuilder;
+use super::PrimitiveBuilder;
 
 /// Array builder for `DictionaryArray`. For example to map a set of byte indices
 /// to f32 values. Note that the use of a `HashMap` here will not scale to very large
@@ -151,9 +174,10 @@ mod tests {
     use super::*;
 
     use crate::array::Array;
-    use crate::bitmap::Bitmap;
-    use crate::buffer::Buffer;
-    use crate::error::Result;
+    use crate::array::UInt32Array;
+    use crate::array::UInt8Array;
+    use crate::datatypes::UInt32Type;
+    use crate::datatypes::UInt8Type;
 
     #[test]
     fn test_primitive_dictionary_builder() {
