+++ arrow/src/array/builder/string_dictionary_builder.rs
@@ -1,9 +1,33 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
 use std::any::Any;
 use std::collections::HashMap;
 use std::sync::Arc;
 
-use crate::array::*;
-use crate::datatypes::*;
+use crate::array::array::Array;
+use crate::array::ArrayBuilder;
+use crate::array::ArrayRef;
+use crate::array::ArrowDictionaryKeyType;
+use crate::array::DictionaryArray;
+use crate::array::PrimitiveBuilder;
+use crate::array::StringArray;
+use crate::array::StringBuilder;
+use crate::datatypes::ArrowNativeType;
 use crate::error::{ArrowError, Result};
 
 /// Array builder for `DictionaryArray` that stores Strings. For example to map a set of byte indices
@@ -199,9 +223,9 @@ mod tests {
     use super::*;
 
     use crate::array::Array;
-    use crate::bitmap::Bitmap;
-    use crate::buffer::Buffer;
-    use crate::error::Result;
+    use crate::array::Int8Array;
+    use crate::datatypes::Int16Type;
+    use crate::datatypes::Int8Type;
 
     #[test]
     fn test_string_dictionary_builder() {
