+++ arrow/src/array/builder/struct_builder.rs
@@ -1,12 +1,28 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
 use std::any::Any;
 use std::fmt;
 use std::sync::Arc;
 
 use crate::array::*;
-use crate::datatypes::*;
+use crate::datatypes::DataType;
+use crate::datatypes::Field;
 use crate::error::Result;
-
-pub use crate::array::builder::GenericBinaryBuilder;
 
 /// Array builder for Struct types.
 ///
@@ -228,7 +244,6 @@ mod tests {
     use crate::array::Array;
     use crate::bitmap::Bitmap;
     use crate::buffer::Buffer;
-    use crate::error::Result;
 
     #[test]
     fn test_struct_array_builder() {
